Develop

See merge request maarch/MaarchCourrier!173

+++ src/app/customField/controllers/CustomFieldController.php
@@ -19,6 +19,7 @@ namespace CustomField\controllers;
 
 use CustomField\models\CustomFieldModel;
 use Group\models\ServiceModel;
+use History\controllers\HistoryController;
 use IndexingModel\models\IndexingModelFieldModel;
 use Respect\Validation\Validator;
 use Slim\Http\Request;
@@ -28,10 +29,6 @@ class CustomFieldController
 {
     public function get(Request $request, Response $response)
     {
-        if (!ServiceModel::hasService(['id' => 'admin_custom_fields', 'userId' => $GLOBALS['userId'], 'location' => 'apps', 'type' => 'admin'])) {
-            return $response->withStatus(403)->withJson(['errors' => 'Service forbidden']);
-        }
-
         $customFields = CustomFieldModel::get();
 
         foreach ($customFields as $key => $customField) {
@@ -65,8 +62,16 @@ class CustomFieldController
         $id = CustomFieldModel::create([
             'label'         => $body['label'],
             'type'          => $body['type'],
-            'values'        => empty($body['values']) ? '[]' : json_encode($body['values']),
-            'default_value' => $body['default_value']
+            'values'        => empty($body['values']) ? '[]' : json_encode($body['values'])
+        ]);
+
+        HistoryController::add([
+            'tableName' => 'custom_fields',
+            'recordId'  => $id,
+            'eventType' => 'ADD',
+            'info'      => _CUSTOMFIELDS_CREATION . " : {$body['label']}",
+            'moduleId'  => 'customField',
+            'eventId'   => 'customFieldCreation',
         ]);
 
         return $response->withStatus(201)->withJson(['customFieldId' => $id]);
@@ -78,9 +83,8 @@ class CustomFieldController
             return $response->withStatus(403)->withJson(['errors' => 'Service forbidden']);
         }
 
-        $field = CustomFieldModel::getById(['select' => [1], 'id' => $args['id']]);
-        if (empty($field)) {
-            return $response->withStatus(400)->withJson(['errors' => 'Custom field not found']);
+        if (!Validator::intVal()->notEmpty()->validate($args['id'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Param id is empty or not an integer']);
         }
 
         $body = $request->getParsedBody();
@@ -91,6 +95,15 @@ class CustomFieldController
             return $response->withStatus(400)->withJson(['errors' => 'Body values is not an array']);
         }
 
+        if (count(array_unique($body['values'])) < count($body['values'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Some values have the same name']);
+        }
+
+        $field = CustomFieldModel::getById(['select' => [1], 'id' => $args['id']]);
+        if (empty($field)) {
+            return $response->withStatus(400)->withJson(['errors' => 'Custom field not found']);
+        }
+
         $fields = CustomFieldModel::get(['select' => [1], 'where' => ['label = ?', 'id != ?'], 'data' => [$body['label'], $args['id']]]);
         if (!empty($fields)) {
             return $response->withStatus(400)->withJson(['errors' => 'Custom field with this label already exists']);
@@ -98,12 +111,20 @@ class CustomFieldController
 
         CustomFieldModel::update([
             'set'   => [
-                'label'         => $body['label'],
-                'values'        => empty($body['values']) ? '[]' : json_encode($body['values']),
-                'default_value' => $body['default_value']
+                'label'  => $body['label'],
+                'values' => empty($body['values']) ? '[]' : json_encode($body['values'])
             ],
             'where' => ['id = ?'],
             'data'  => [$args['id']]
+        ]);
+
+        HistoryController::add([
+            'tableName' => 'custom_fields',
+            'recordId'  => $args['id'],
+            'eventType' => 'UP',
+            'info'      => _CUSTOMFIELDS_MODIFICATION . " : {$body['label']}",
+            'moduleId'  => 'customField',
+            'eventId'   => 'customFieldModification',
         ]);
 
         return $response->withStatus(204);
@@ -115,13 +136,28 @@ class CustomFieldController
             return $response->withStatus(403)->withJson(['errors' => 'Service forbidden']);
         }
 
-        IndexingModelFieldModel::delete(['where' => ['type = ?', 'identifier = ?'], 'data' => ['custom', $args['id']]]);
+        if (!Validator::intVal()->notEmpty()->validate($args['id'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Param id is empty or not an integer']);
+        }
+
+        $field = CustomFieldModel::getById(['select' => ['label'], 'id' => $args['id']]);
+
+        IndexingModelFieldModel::delete(['where' => ['identifier = ?'], 'data' => [$args['id']]]);
 
         //TODO Suppression des valeurs liés aux courriers ?
 
         CustomFieldModel::delete([
             'where' => ['id = ?'],
             'data'  => [$args['id']]
+        ]);
+
+        HistoryController::add([
+            'tableName' => 'custom_fields',
+            'recordId'  => $args['id'],
+            'eventType' => 'DEL',
+            'info'      => _CUSTOMFIELDS_SUPPRESSION . " : {$field['label']}",
+            'moduleId'  => 'customField',
+            'eventId'   => 'customFieldSuppression',
         ]);
 
         return $response->withStatus(204);
